linux: refresh libraries to depend on libscigcc_s.so

+++ build-linux-dependencies.sh
@@ -158,11 +158,16 @@ build_gcc() {
 	# rename GCC library names to SCI ones to avoid dependency on system ones
 	patchelf --set-soname libsciquadmath.so.0 $INSTALLDIR$LIBDIR/libquadmath.so.0.0.0
 	cp -a $INSTALLDIR$LIBDIR/libquadmath.so.0.0.0 $INSTALLDIR/lib/libsciquadmath.so.0
+	patchelf --set-soname libscigcc_s.so.1 $INSTALLDIR$LIBDIR/libgcc_s.so.1
+	cp -a $INSTALLDIR$LIBDIR/libgcc_s.so.1 $INSTALLDIR/lib/libscigcc_s.so.1
 	patchelf --set-soname libscigfortran.so.5 $INSTALLDIR$LIBDIR/libgfortran.so.5.0.0
 	patchelf --replace-needed libquadmath.so.0 libsciquadmath.so.0 $INSTALLDIR$LIBDIR/libgfortran.so.5.0.0
-	patchelf --remove-rpath $INSTALLDIR$LIBDIR/libgfortran.so.5.0.0
+	patchelf --replace-needed libgcc_s.so.1 libscigcc_s.so.1 $INSTALLDIR$LIBDIR/libgfortran.so.5.0.0
+        patchelf --remove-rpath $INSTALLDIR$LIBDIR/libgfortran.so.5.0.0
 	cp -a $INSTALLDIR$LIBDIR/libgfortran.so.5.0.0 $INSTALLDIR/lib/libscigfortran.so.5
 	patchelf --set-soname libscistdc++.so.6 $INSTALLDIR$LIBDIR/libstdc++.so.6.0.25
+	patchelf --replace-needed libgcc_s.so.1 libscigcc_s.so.1 $INSTALLDIR$LIBDIR/libstdc++.so.6.0.25
+        patchelf --remove-rpath $INSTALLDIR$LIBDIR/libstdc++.so.6.0.25
 	cp -a $INSTALLDIR$LIBDIR/libstdc++.so.6.0.25 $INSTALLDIR/lib/libscistdc++.so.6
 	
         cd ../..
@@ -746,8 +751,11 @@ do
       rm -f $LIBTHIRDPARTYDIR/libcamd.*
       cp -d $INSTALLDIR/lib/libcamd.* $LIBTHIRDPARTYDIR/
 
-      rm -f $LIBTHIRDPARTYDIR/libOpenXLSX.*
-      cp -d $INSTALLDIR/lib/libOpenXLSX.* $LIBTHIRDPARTYDIR/
+      if stat -t $INSTALLDIR/lib/libOpenXLSX.* >/dev/null 2>&1
+      then
+          rm -f $LIBTHIRDPARTYDIR/libOpenXLSX.*
+          cp -d $INSTALLDIR/lib/libOpenXLSX.* $LIBTHIRDPARTYDIR/
+      fi
 
       # Scilab dependencies where the system ones are not recent enough to be used.
       #
